intention_analyzer/pyproject.toml
[tool.poetry]
name = "intention_analyzer"
version = "0.1.0"
description = "A tool for analyzing user intentions in natural language."
authors = ["Sentinel Being <@cisco.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.0,<4.0"
langchain = "^0.3.17"
langgraph = "^0.3.5"
langchain-openai = "^0.3.3"
jinja2 = "^3.1.5"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.10.0"
sentence-transformers = "^4.0.2"

[tool.poetry.group.manifest.dependencies]
agntcy_acp = "v0.1.0a2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
